Remove passions, age group interests, and skills models
The hard defined opportunity categories are replaced with tags and tag categories. This is a hard change and not elegant. The migration deletes all of the passion, age group interest, and skill entries. No one else is using this app as far as I am aware so this skips on attempting to migrate the old category entries to tags.