Hard restart of migration history
This is a hard restart of the migration history that would likely break any existing database. Normally this is not a possible option, but no one has forked this project yet. This choice was made rather than spending time on a migration error. I bring shame on my family :P