Enhance and replace authentication with JWT token authentication
Flask httpauth is now only used for the initial basic user authentication. Once the user logs in a JWT access token is generated and given a expiration time. A refresh token is also generated that is used to generate a new access token. All protected routes are now protected with jwt_required.

The access tokens are stored in httponly cookies. This will be protected by https in production but needs further testing. There is also CSRF protection available but was causing issues so will have to be figured outbefore ready for production.