Replace broken JWT auth with working HTTPbasic and HTTPtoken auth
This removes the flask-jwt-extended package and replaces all of the authentication functionality with Flask-HTTPauth. There are a lot of changes here but it is largely replacing the JWT decorators with HTTPauth decorators.

Changes here do include incorporating user role requirements for some endpoints. This is to protect base level setup management and allow access only to users with the Admin role.