Remove Frequency model and add Opportunity frequency fields
The Frequency model and relationship field in Opportunity is replaced with two frequency fields: frequency_unit and frequency_modifier. The frequency_unit is meant to specify the frequency unit such as Day, Week, Month, Thursday. The frequency_modifier is meany to a modification of the unit such as 1st, 2nd, 3rd. With the two fields the frequency can be set as every 2nd Week or every 3rd Wednesday.

This includes the api portion of ticket #40